use unint64_t to sort read/write fields

+++ src/util/sort_fields.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #include "../process.h"
 #include "../../display/display.h"
 
@@ -5,4 +7,4 @@
 proc_t *sort_by_field(proc_t *procs, int field, int nproc);
 
 void cur_fields(proc_t *proc_arr[], proc_t *cur, int proc_index,
-                int field, int cmp_fields[]);
+                int field, uint64_t cmp_fields[]);
